stress-inode-flags: flags argument should be long int rather than int
valgrind detected that the ioctl flags arguement was not fully initialized
because an int was being used and not a long int. Fix this.

+++ stress-inode-flags.c
@@ -82,7 +82,7 @@ static void stress_inode_flags_ioctl(
 static inline void stress_inode_flags_ioctl_sane(const int fd)
 {
 	int ret;
-	const int flag = 0;
+	const long int flag = 0;
 
 	ret = ioctl(fd, FS_IOC_SETFLAGS, &flag);
 	(void)ret;
